@@ -7,11 +7,12 @@ namespace Task2Infrastructure.EntityFramework
     [UsedImplicitly]
     public sealed class BooksContextDesignTimeFactory : IDesignTimeDbContextFactory<BooksContext>
     {
-        private const string DefaultConnectionString = "Data Source=127.0.0.1;Initial Catalog=EFTestApplication;User Id=sa; Password=2wsx2WSX;";
+        //private const string DefaultConnectionString = "Data Source=127.0.0.1;Initial Catalog=BookShop;User Id=postgres; Password=NSU;";
+        private const string DefaultConnectionString = "Host=localhost;Username=postgres;Password=NSU;Database=BookShop;";
         public static DbContextOptions<BooksContext> GetSqlServerOptions([CanBeNull]string connectionString)
         {
             return new DbContextOptionsBuilder<BooksContext>()
-                .UseSqlServer(connectionString ?? DefaultConnectionString, x =>
+                .UseNpgsql(connectionString ?? DefaultConnectionString, x =>
                 {
                     x.MigrationsHistoryTable("__EFMigrationsHistory", BooksContext.DefaultSchemaName);
                 })
