@@ -19,20 +19,16 @@ namespace Task2WebApplication.Controllers
         [HttpGet]
         public async Task<List<Book>> GetBooks()
         {
-            using (var context = _dbContextFactory.GetContext())
-            {
-                return await context.GetBooks();
-            }
+            await using var context = _dbContextFactory.GetContext();
+            return await context.GetBooks();
         }
 
         [HttpPost]
         public async Task AddBook([FromBody] Book book)
         {
-            using (var context = _dbContextFactory.GetContext())
-            {
-                context.AddBook(book);
-                await context.SaveChangesAsync();
-            }
+            await using var context = _dbContextFactory.GetContext();
+            context.AddBook(book);
+            await context.SaveChangesAsync();
         }
     }
 }
